back/requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
audiocraft==1.3.0
flask>=3.0.2
flask-cors>=4.0.0
diffusers>=0.25.0
transformers>=4.36.2
accelerate>=0.27.0
xformers==0.0.22.post7
huggingface-hub>=0.20.0
safetensors>=0.4.0
invisible-watermark>=0.2.0
scipy>=1.12.0
gunicorn>=21.2.0
nvidia-ml-py>=12.535.77
pillow>=10.0.0